\name{runLicorMatrix}
\alias{runLicorMatrix}
\title{Presents the packages graphical user interface}
\usage{
  runLicorMatrix(port = 1972L)
}
\arguments{
  \item{port}{the port where to listen; 1972 by default.}
}
\description{
  Runs a web server to show the user interface.
}
\author{
  Reinhard Simon
}

